requirements.txt
numpy
scikit-learn
pyyaml
jupyter
matplotlib
seaborn